requirements.txt
redis==2.10.3
msgpack-python==0.4.4
